+++ cypress/e2e/marc/marc-bibliographic/create-new-marc-bib/search-created-marc-bib-record-by-title-OCLC-number.cy.js
@@ -37,7 +37,9 @@ describe('MARC', () => {
       });
       after('Delete test data', () => {
         cy.getAdminToken();
-        Users.deleteViaApi(userId);
+        if (userId) {
+          Users.deleteViaApi(userId);
+        }
         InventoryInstance.deleteInstanceViaApi(createdInstanceIDs[0]);
       });
